clarify uid/gid mapping entry format
The manual is extremely unclear (counter-intuitive) on the ID mapping
format. The "-o idmap=TYPE" language contains "[h]ow to map remote
UID/GIDs to local values", suggesting the direction remote -> local. But
then the entry format in the "-o uidfile" and "-o gidfile" files is
precisely the opposite. The command line help hints at this, but the
manual is entirely obscure. Clarify both locations.

Signed-off-by: Laszlo Ersek <lersek@redhat.com>